pyproject.toml
[tool.poetry]
name = "atlassian-access-user-audit"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
uplink = "^0.9.4"
furl = "^2.1.2"
python-dateutil = "^2.8.2"
fire = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
